complaint_system/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autobahn==23.6.2
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.0.6
django-event==1.0.3
django-widget-tweaks==1.5.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.0
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.7
hyperlink==21.0.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_proxy==4.3.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_vpython==3.1.8
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.1
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.9.0
PyOpenGL==3.1.7
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
razorpay==1.4.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.5
scipy==1.14.0
screen_brightness_control==0.23.0
Send2Trash==1.8.3
setuptools==71.1.0
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.1
sockjs-tornado==1.0.7
sounddevice==0.4.7
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.13.2
terminado==0.18.1
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
vpython==7.6.5
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
WMI==1.5.1
yarl==1.9.4
